Revert "ref: begin python3 conversion"
This reverts commit 176c216d59b6ecedec498ce7f70ee87f1a3e06d4.

+++ server/HLWebServices.py
@@ -1,9 +1,8 @@
-from __future__ import absolute_import
 from twisted.web import xmlrpc , server
 from twisted.internet import reactor
 from shared.HLTypes import *
 from config import *
-from six.moves.xmlrpc_client import Binary
+from xmlrpclib import Binary
 import time
 
 class HLWebServices( xmlrpc.XMLRPC ):
@@ -17,7 +16,7 @@ class HLWebServices( xmlrpc.XMLRPC ):
 	
 	def xmlrpc_getServerUptime( self ):
 		""" Returns the server uptime in seconds. """
-		return int( time.time() - self.server.startTime )
+		return long( time.time() - self.server.startTime )
 	
 	def xmlrpc_getUserlist( self ):
 		""" Returns a list of online users. Each entry is a dictionary containing user information. """
